implementação dialog

+++ src/app/materia/materia.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { DataStoreService } from '../data-storage';
 import { Materia, MateriaPayload } from '../entities/materia.entity';
+import { DialogService } from '../services/dialog.service';
 import { MateriaService } from '../services/materia.service';
 
 @Component({
@@ -18,7 +19,8 @@ export class MateriaComponent implements OnInit {
   constructor(
     private materiaService: MateriaService,
     private snackBar: MatSnackBar,
-    private dataStorage: DataStoreService
+    private dataStorage: DataStoreService,
+    private dialogService: DialogService
   ) {
     this.dataStorage.isSmall.subscribe((e) => (this.isSmall = e));
   }
@@ -58,14 +60,21 @@ export class MateriaComponent implements OnInit {
   }
 
   excluir(materia: Materia) {
-    this.materiaService.deleteMateria(materia.id).subscribe({
-      next: (res) => {
-        this.snackBar.open(res.msg, undefined, { duration: 5000 });
-        this.search();
-      },
-      error: (err) => {
-        this.snackBar.open(err.error.msg, undefined, { duration: 5000 });
-      }
-    });
+    this.dialogService
+      .openConfirmDialog()
+      .afterClosed()
+      .subscribe((res1) => {
+        if (res1) {
+          this.materiaService.deleteMateria(materia.id).subscribe({
+            next: (res) => {
+              this.snackBar.open(res.msg, undefined, { duration: 5000 });
+              this.search();
+            },
+            error: (err) => {
+              this.snackBar.open(err.error.msg, undefined, { duration: 5000 });
+            }
+          });
+        }
+      });
   }
 }
